fix(SWS_Biofoam): Updated paths for Biofoam to fix recent error (#299)
* Fix Ammo in Elite Teamlead

Gave the Elite Teamlead the correct ammo and removed certain heads that conflict with headwear.

* Update CfgWeapons.inc

* Update CfgWeapons.inc

* Update CfgWeapons.inc

Updated the paths for the model and image for biofoam

* Updates SM247 & Reaper Autorilfe

Changed paths for the SM247s model, hidden selections, and magazine icons. Updated the Reaper Autorifle's weapon to the SM73 due to the removal ol the DMNS Confetti Maker